api_dev/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contextlib2==21.6.0
cryptography==44.0.1
daphne==4.2.1
distlib==0.3.8
Django==5.2.5
django-axes==7.0.2
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-extensions==3.2.3
django-fernet-fields==0.6
django-model-utils==5.0.0
django-ninja==1.4.3
django-ninja-jwt==5.3.5
django-otp==1.5.4
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.13.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
injector==0.22.0
kombu==5.4.2
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.7
pydantic_core==2.33.2
pyOpenSSL==25.0.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
zope.interface==7.2
Pillow>=10.0.0

